dmp 2537 more response to comments
