+++ src/functionalTest/java/uk/gov/hmcts/darts/usermanagement/SecurityGroupFunctionalTest.java
@@ -14,10 +14,13 @@ import org.skyscreamer.jsonassert.RegularExpressionValueMatcher;
 import org.skyscreamer.jsonassert.comparator.CustomComparator;
 import uk.gov.hmcts.darts.FunctionalTest;
 import uk.gov.hmcts.darts.common.enums.SecurityRoleEnum;
+import uk.gov.hmcts.darts.courthouse.model.ExtendedCourthousePost;
 import uk.gov.hmcts.darts.usermanagement.model.SecurityGroupWithIdAndRole;
+import uk.gov.hmcts.darts.usermanagement.model.SecurityGroupWithIdAndRoleAndUsers;
 
 import java.util.Comparator;
 import java.util.List;
+import java.util.UUID;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.Matchers.hasSize;
@@ -137,5 +140,100 @@ class SecurityGroupFunctionalTest extends FunctionalTest {
         if (courtroomId != null) {
             assertTrue(group.getCourthouseIds().contains(courtroomId));
         }
+    }
+
+    @Test
+    void shouldPatchSecurityGroups() throws JsonProcessingException {
+
+        String postContent = """
+                         {
+                           "name": "<func-a-security-group>",
+                           "display_name": "<A security group>",
+                           "description": "func-test group"
+                         }
+                           """;
+        postContent = postContent.replace("<func-a-security-group>", "func-a-security-group " + UUID.randomUUID());
+        postContent = postContent.replace("<A security group>", "A security group " + UUID.randomUUID());
+
+        Response response = buildRequestWithExternalGlobalAccessAuth()
+            .baseUri(getUri("/admin/security-groups"))
+            .contentType(ContentType.JSON)
+            .body(postContent)
+            .post()
+            .thenReturn();
+
+        assertEquals(201, response.getStatusCode());
+
+        String patchContent = """
+                         {
+                           "name": "<func-a-security-group-new-name>",
+                           "display_name": "<A security group new name>",
+                           "description": "func-test group new description",
+                           "courthouse_ids": [<id1>,<id2>],
+                           "user_ids": [-1,-2]
+                         }
+                           """;
+        String newName = "func-a-security-group-new-name " + UUID.randomUUID();
+        patchContent = patchContent.replace("<func-a-security-group-new-name>", newName);
+        String newDisplayName = "A security group new name " + UUID.randomUUID();
+        patchContent = patchContent.replace("<A security group new name>", newDisplayName);
+        Integer id1 = createCourthouse("func-a_courthouse " + UUID.randomUUID(), "func-a_courthouse" + UUID.randomUUID());
+        Integer id2 = createCourthouse("func-a_courthouse " + UUID.randomUUID(), "func-a_courthouse" + UUID.randomUUID());
+        patchContent = patchContent.replace("<id1>", id1.toString());
+        patchContent = patchContent.replace("<id2>", id2.toString());
+
+        SecurityGroupWithIdAndRole securityGroupWithIdAndRoles = MAPPER.readValue(response.asString(),
+                                                                                  new TypeReference<SecurityGroupWithIdAndRole>(){});
+
+        response = buildRequestWithExternalGlobalAccessAuth()
+            .baseUri(getUri("/admin/security-groups/" + securityGroupWithIdAndRoles.getId()))
+            .contentType(ContentType.JSON)
+            .body(patchContent)
+            .patch()
+            .thenReturn();
+
+        SecurityGroupWithIdAndRoleAndUsers securityGroupWithIdAndRoleAndUsers = MAPPER.readValue(response.asString(),
+                                                                                                 new TypeReference<SecurityGroupWithIdAndRoleAndUsers>(){});
+
+        assertEquals(newName, securityGroupWithIdAndRoleAndUsers.getName());
+        assertEquals(newDisplayName, securityGroupWithIdAndRoleAndUsers.getDisplayName());
+        assertEquals("func-test group new description", securityGroupWithIdAndRoleAndUsers.getDescription());
+        assertTrue(securityGroupWithIdAndRoleAndUsers.getCourthouseIds().contains(id1));
+        assertTrue(securityGroupWithIdAndRoleAndUsers.getCourthouseIds().contains(id2));
+        assertTrue(securityGroupWithIdAndRoleAndUsers.getUserIds().contains(-1));
+        assertTrue(securityGroupWithIdAndRoleAndUsers.getUserIds().contains(-2));
+
+        response = buildRequestWithExternalGlobalAccessAuth()
+            .baseUri(getUri("/admin/security-groups/" + securityGroupWithIdAndRoles.getId()))
+            .contentType(ContentType.JSON)
+            .get()
+            .thenReturn();
+
+        SecurityGroupWithIdAndRole retrievedSecurityGroupWithIdAndRoles = MAPPER.readValue(response.asString(),
+                                                                                           new TypeReference<SecurityGroupWithIdAndRole>(){});
+
+        assertEquals(newName, retrievedSecurityGroupWithIdAndRoles.getName());
+        assertEquals(newDisplayName, retrievedSecurityGroupWithIdAndRoles.getDisplayName());
+        assertEquals("func-test group new description", retrievedSecurityGroupWithIdAndRoles.getDescription());
+        assertTrue(retrievedSecurityGroupWithIdAndRoles.getCourthouseIds().contains(id1));
+        assertTrue(retrievedSecurityGroupWithIdAndRoles.getCourthouseIds().contains(id2));
+        assertTrue(retrievedSecurityGroupWithIdAndRoles.getUserIds().contains(-1));
+        assertTrue(retrievedSecurityGroupWithIdAndRoles.getUserIds().contains(-2));
+
+
+    }
+
+    private Integer createCourthouse(String name, String displayName) throws JsonProcessingException {
+        String content = String.format("{\"courthouse_name\": \"%s\", \"display_name\": \"%s\"}", name, displayName);
+        Response response = buildRequestWithExternalGlobalAccessAuth()
+            .baseUri(getUri("/admin/courthouses"))
+            .contentType(ContentType.JSON)
+            .body(content)
+            .post()
+            .thenReturn();
+
+        ExtendedCourthousePost extendedCourthousePost = MAPPER.readValue(response.asString(),
+                                                                         new TypeReference<ExtendedCourthousePost>(){});
+        return extendedCourthousePost.getId();
     }
 }
